+++ Quilt/src/main/java/me/ultrusmods/missingwilds/client/MissingWildsClientQuilt.java
@@ -3,7 +3,9 @@ package me.ultrusmods.missingwilds.client;
 import me.ultrusmods.missingwilds.Constants;
 import me.ultrusmods.missingwilds.MissingWildsQuilt;
 import me.ultrusmods.missingwilds.client.render.FireflySwarmRenderer;
+import me.ultrusmods.missingwilds.compat.template.TemplateModCompatClient;
 import me.ultrusmods.missingwilds.particle.FireflyParticle;
+import me.ultrusmods.missingwilds.platform.Services;
 import me.ultrusmods.missingwilds.register.MissingWildsEntities;
 import me.ultrusmods.missingwilds.register.MissingWildsParticles;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
@@ -29,6 +31,8 @@ public class MissingWildsClientQuilt implements ClientModInitializer {
 
 //        ColorProviderRegistry.BLOCK.register((blockState, blockAndTintGetter, blockPos, col) -> blockAndTintGetter != null && blockPos != null ? 2129968 : 7455580, MissingWildsBlocks.WATERLILY_BLOCK.get());
 //        ColorProviderRegistry.ITEM.register((itemStack, col) -> 7455580, MissingWildsItems.WATERLILY_ITEM.get());
-
+        if (Services.PLATFORM.isModLoaded("templates")) {
+            TemplateModCompatClient.init();
+        }
     }
 }